Added file logger. Added logs in match case. Fixed fire stations generation. Started implementation of control panel.

+++ project/src/main/scala/pcd/assignment03/distributed_programming/model/Boundary.scala
@@ -7,30 +7,30 @@ trait Boundary:
   /**
    * left bound
    */
-  def x0: Double
+  def x0: Int
   /**
    * top bound
    */
-  def y0: Double
+  def y0: Int
   /**
    * right bound
    */
-  def x1: Double
+  def x1: Int
   /**
    * bottom bound
    */
-  def y1: Double
+  def y1: Int
   
-  def width: Double = x1 - x0
+  def width: Int = x1 - x0
   
-  def height: Double = y1 - y0
+  def height: Int = y1 - y0
 
 object Boundary:
-  def apply(x0: Double, y0: Double, x1: Double, y1: Double): Boundary = BoundaryImpl(x0, y0, x1, y1)
+  def apply(x0: Int, y0: Int, x1: Int, y1: Int): Boundary = BoundaryImpl(x0, y0, x1, y1)
 
   private case class BoundaryImpl(
-                                   override val x0: Double,
-                                   override val y0: Double,
-                                   override val x1: Double,
-                                   override val y1: Double
+                                   override val x0: Int,
+                                   override val y0: Int,
+                                   override val x1: Int,
+                                   override val y1: Int
                                  ) extends Boundary
